Inline W4A16 benchmark helpers
